pyproject.toml
[project]
name = "spim_core"
description = "Core features common across Allen Institute spim systems."
version = "0.1.0"
authors = [
    {name = "Adam Glaser", email = "adam.glaser@alleninstitute.org"},
    {name = "Joshua Vasquez", email = "joshua.vasquez@alleninstitute.org"}
]
readme = "README.md"
keywords = ['microcsopy', 'spim']
classifiers=[
    'License :: OSI Approved :: MIT License',
    'Development Status :: 4 - Beta',
    'Programming Language :: Python'
]
#packages = [
#    { include = "spim_core"},
#]

requires-python = ">=3.8"
dependencies = [
    "PyImarisWriter >= 0.7.0",
    "thorlabs-apt >= 0.2",
    "tigerasi @ git+ssh://git@github.com/AllenNeuralDynamics/TigerASI.git@main#egg=TigerASI",
    "nidaqmx >= 0.6.2",
    "numpy >= 1.22.3",
    "matplotlib >= 3.5.2",
    "toml >= 0.10.2",
    "mock >= 4.0.3",
    "coloredlogs >= 15.0.1",
    "gitPython >= 3.1.27",
]

[project.urls]
repository = "https://github.com/AllenNeuralDynamics/exa-spim-control"

[build-system]
requires = ["flit-core>=3.7.1,<4"]
build-backend = "flit_core.buildapi"

[project.scripts]
exaspim = "bin.main:main"
